PI-744 Script to re-organise secrets
This script re-organises secrets from the flat structure we have at the moment (due to them being synced from GitHub), into separate secrets for different sets of credentials. E.g. instead of all values being in a single `approved-premises-and-delius` secret, they'll be separated into `approved-premises-and-delius-database`, `approved-premises-and-delius-queue`, `approved-premises-and-delius-client-credentials` etc.

This enables each secret to be individually managed - so the queue/topic secrets can be managed by Terraform, the OAuth credentials can be managed by the auth team, etc.